requirements.txt
azure-functions
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.5
idna==3.9
pytz==2024.2
requests==2.32.3
setuptools==74.1.2
soupsieve==2.6
urllib3==2.2.3
zope.interface==7.0.3
